improve test suite speed
