+++ src/test/kotlin/uk/gov/justice/digital/hmpps/assessments_events/utils/LastAccessedEventHelperTest.kt
@@ -2,16 +2,9 @@ package uk.gov.justice.digital.hmpps.assessments_events.utils
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.ActiveProfiles
+import uk.gov.justice.digital.hmpps.assessments_events.integration.IntegrationTestBase
 
-@SpringBootTest
-@ActiveProfiles("test")
-class LastAccessedEventHelperTest {
-
-  @Autowired
-  lateinit var lastAccessedEvent: LastAccessedEventHelper
+class LastAccessedEventHelperTest : IntegrationTestBase() {
 
   @Test
   fun `Can read Last Accessed Event date from properties`() {
